# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "styles": {
            "items": {
              "$ref": "#/components/schemas/StyleOut"
            },
            "title": "Styles",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "styles"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StyleOut": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "cost": {
            "title": "Cost",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "examples": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Examples",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_spec": {
            "title": "Input Spec",
            "type": "object"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "operation",
          "input_spec",
          "cost"
        ],
        "title": "StyleOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Identity, wallet and AI content generation for TOONTOON.",
    "title": "ARTEKI API",
    "version": "generated"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/shots/daily": {
      "get": {
        "description": "Today's set. Same for everyone, reproducible for any past date.",
        "operationId": "daily_shots_api_shots_daily_get",
        "parameters": [
          {
            "description": "YYYY-MM-DD (UTC), по умолчанию сегодня",
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM-DD (UTC), по умолчанию сегодня",
              "title": "Day"
            }
          },
          {
            "in": "query",
            "name": "count",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 12,
              "minimum": 1,
              "title": "Count",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StyleOut"
                  },
                  "title": "Response Daily Shots Api Shots Daily Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Daily Shots",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/styles": {
      "get": {
        "description": "The whole catalogue, grouped by direction.\n\nПорядок разделов — тот, что выставлен вручную в `CATEGORIES`, и он один\nдля всех. Внутри раздела — тоже заданный.\n\nПерсонализация (♥ из онбординга наверх) написана и лежит рядом, но\nвыключена флагом `personalise_catalogue`. Она переворачивала верх витрины:\nчеловек, отметивший шесть направлений из одиннадцати, видел первыми их, и\nвыстроенный порядок начинал работать только с седьмой ленты. Пока задача —\nпервое впечатление, а оно должно быть одинаковым и предсказуемым.",
        "operationId": "list_styles_api_styles_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "arteki-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arteki-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryOut"
                  },
                  "title": "Response List Styles Api Styles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Styles",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/styles/home": {
      "get": {
        "description": "The showcase: a few strong examples, not a grid of everything (CH-14).\n\nРанжирование под человека — тем же флагом, что и порядок разделов: иначе\nленты шли бы одинаково у всех, а витрина над ними — по-разному, и первое\nвпечатление опять зависело бы от ответов в онбординге.",
        "operationId": "home_styles_api_styles_home_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 24,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "arteki-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arteki-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StyleOut"
                  },
                  "title": "Response Home Styles Api Styles Home Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Home Styles",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/styles/{style_id}": {
      "get": {
        "operationId": "get_style_api_styles__style_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "style_id",
            "required": true,
            "schema": {
              "title": "Style Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Style",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/styles/{style_id}/example/{index}": {
      "get": {
        "description": "Витринная картинка стиля — публично и с длинным кэшем.\n\nЭто единственная картинка в системе, которую можно отдавать кому угодно:\nона наша, снята для каталога и показывается до входа. Всё остальное живёт\nза `/api/media` с проверкой владельца.",
        "operationId": "style_example_api_styles__style_id__example__index__get",
        "parameters": [
          {
            "in": "path",
            "name": "style_id",
            "required": true,
            "schema": {
              "title": "Style Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "schema": {
              "title": "Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Style Example",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}